backend/app/requirements.txt
fastapi
uvicorn[standard]
motor
python-dotenv
python-jose
passlib[bcrypt]
